requirement.txt
django-ninja
Django
mysqlclient
django-cors-headers
psycopg2